nlp-module/requirements.txt
flask==2.0.2
googletrans-4.0.0rc1